pyproject.toml
[project]
name = "gohighlevel-mcp"
version = "0.1.0"
description = "MCP server for GoHighLevel API integration"
dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "python-telegram-bot>=20.7"
]

[project.scripts]
gohighlevel-mcp = "gohighlevel_mcp.server:main"
telegram-bot = "telegram_bot:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0"
]